Add HAZUS hurricane library files and generation code.
